Beta parameters from interpretable conditions

Rather than knowing intrinsically what parameters are required for a distribution, scientists tend to have a sense of what value they expect a measure to take, how much they expect observations to fall within a certain distance of that value.

beta_pal(expected_value, within, this_much)

Arguments

expected_value
Expected value of beta distrbution from [0,1].

within
Specify distance this_much falls within.

this_much
What proportion falls within the specified interval.

Value

List of parameters appropriate for rbeta family of functions.

Details

This solution coded by by this most helpful gist.

Visualise the distribution with beta_plot. See vignette("beta_pal") for derivations and more information.